separate bindings into fumenutil and fumen-editor
globals are pending to be fixed
