+++ index.html
@@ -13,12 +13,12 @@
 <!--EXTERNAL SCRIPTS-->
 <script type = "module" src = "./base.js"></script>
 <script src = "./Fumen Utils_files/tetris-fumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-glueFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-unglueFumen.js"></script>
+<!-- <script type = "module" src = "./Fumen Utils_files/modified-glueFumen.js"></script> -->
+<!-- <script type = "module" src = "./Fumen Utils_files/modified-unglueFumen.js"></script>
 <script type = "module" src = "./Fumen Utils_files/modified-mirrorFumen.js"></script>
 <script type = "module" src = "./Fumen Utils_files/modified-combineFumen.js"></script>
 <script type = "module" src = "./Fumen Utils_files/modified-splitFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-removeComments.js"></script>
+<script type = "module" src = "./Fumen Utils_files/modified-removeComments.js"></script> -->
 <!-- <script type = "module" src = "./Fumen Utils_files/modified-fumen-canvas.js"></script>
 <script type = "module" src = "./Fumen Utils_files/modified-fumen-render.js"></script> -->
 <!-- <script src = "./Fumen Utils_files/board-render.js"></script> -->
@@ -61,7 +61,7 @@
 			<div class = "fumenOperations">
 				<div class="tooltip">
 					<button id="clear-input" class = "ibtn">Clear Input</button>
-					<span class="tooltiptext">Shortcut: Backspace, Cancel Action: Escape</span>
+					<span class="tooltiptext">Shortcut: Backspace, Cancel: Escape</span>
 				</div>
 				<div class="tooltip">
 					<button id="glue-fumen" class = "ibtn"><u>G</u>lue</button>
@@ -334,7 +334,7 @@
 								</div>	
 								<div class = "tooltip">
 									<button id = "duplicatePage" class = "fumen-option"><b>Duplicate Page</b></button>
-									<span class = "tooltiptext">Duplicate the current page excluding mino mode. Shortcut: D</span>
+									<span class = "tooltiptext">Duplicate the current page excluding mino mode. Shortcut: Shift+D, P</span>
 								</div>
 								<div class = "tooltip">
 									<button id = "clearPage" class = "fumen-option"><b>Clear Page</b></button>
@@ -346,7 +346,7 @@
 								</div>
 								<div class = "tooltip">
 									<button id = "reset" class="fumen-option"><b>Reset</b></button>
-									<span class = "tooltiptext">Delete all data. Shortcut: R, Cancel Action: Escape</span>
+									<span class = "tooltiptext">Delete all data. Shortcut: R, Cancel: Escape</span>
 								</div><hr class="small-hr">
 								<div class = "tooltip">
 									<button id = "importImage" class = "fumen-option"><b>Import Image</b></button>
